refactor(test): 인증 및 사용자 테스트 구조 개선
기존에 JWT 관련 파일에 혼재되어 있던 인증/인가 테스트를 분리하여 `auth` 폴더를 신설하고, 새로운 인증 테스트와 함께 통합했습니다. 이를 통해 인증 관련 테스트의 응집도를 높였습니다.

또한, 흩어져 있던 사용자 관련 테스트들을 `user` 폴더로 옮겨 명확하게 분리하고 관리의 효율성을 개선했습니다.
